api/requirements.txt
Flask==1.1.2
psycopg2-binary
pydantic